added the takeHit(int damageTaken) function that will allow for the player to take damage from hits, and will return a bool stating if the ship is still alive.
updated afterboosterActive_() function, to now use afterburnerSpeed_ variable that is in each ship class, also reduced booster speed to 5 from 1000 so that the ship does not disappear when you hit spacebar.

noticed that the bankLeft() and bankRight() functions do not take currentSpeed into account when translating the ship left or right.

changed error messages for when bitmaps fail to load from files, for the background, singleplayer, and multiplayer images. It is now more useful, it states which image failed not just that a bitmap failed to load.